исправил переменные в map.js и добавил функцию для генерации карточки в popup.js
